remove mock module
