+++ js/main.js
@@ -1,4 +1,3 @@
-import './mock.js';
 import './popup.js';
 import './validity.js';
 import './map.js';
