+++ js/popup.js
@@ -1,4 +1,25 @@
-import {TYPES} from './mock.js';
+const TYPES = {
+  palace: {
+    name: 'Дворец',
+    minPrice: 10000,
+  },
+  flat: {
+    name: 'Квартира',
+    minPrice: 1000,
+  },
+  house: {
+    name: 'Дом',
+    minPrice: 5000,
+  },
+  bungalow: {
+    name: 'Бунгало',
+    minPrice: 0,
+  },
+  hotel: {
+    name: 'Отель',
+    minPrice: 3000,
+  },
+};
 
 const createCustomPopup = ({author, offer}) => {
   const {avatar} = author;
@@ -76,4 +97,4 @@ const createCustomPopup = ({author, offer}) => {
   return offerElement;
 };
 
-export {createCustomPopup};
+export {createCustomPopup, TYPES};
